Moved `strategy` into subfolder and added `estimated_arm_sds`
