+++ strategy/__init__.py
@@ -1,0 +1,2 @@
+from strategy.strategy import *
+from strategy.thompson_sampling_strategy import *
