+++ thompson_sampling_script.py
@@ -22,7 +22,7 @@ if __name__ == '__main__':
     print('Prior params: ' + str(bernoulli_strategy.prior_params))
     print('Posterior params: ' + str(bernoulli_strategy.posterior_params))
     print('True means: ' + str([0.45, 0.55]))
-    print('Mean reward estimates: ' + str(bernoulli_strategy.mean_reward_estimates))
+    print('Arm mean reward estimates: ' + str(bernoulli_strategy.estimated_arm_means))
 
     # Gaussian
     gaussian_bandit = StaticBandit(arms=[GaussianArm(mu=95, sigma=30),
@@ -34,7 +34,7 @@ if __name__ == '__main__':
     print('Prior params: ' + str(gaussian_strategy.prior_params))
     print('Posterior params: ' + str(gaussian_strategy.posterior_params))
     print('True means: ' + str([95, 105]))
-    print('Mean reward estimates: ' + str(gaussian_strategy.mean_reward_estimates))
+    print('Arm mean reward estimates: ' + str(gaussian_strategy.estimated_arm_means))
 
     # Linear interpolation bandit with Gaussian errors
     dynamic_bandit = LinearInterpolationBandit(means=np.array([[5.0, 8.0], [10.0, 5.0]]),
@@ -43,7 +43,7 @@ if __name__ == '__main__':
                                                   sigma=20,
                                                   mu_prior=0, sigma_prior=200,
                                                   memory_multiplier=0.9)
-    dynamic_strategy.fit(iterations=1000, plot=True)
+    out = dynamic_strategy.fit(iterations=1000)
     print('Prior params: ' + str(dynamic_strategy.prior_params))
     print('Posterior params: ' + str(dynamic_strategy.posterior_params))
